refactor: #108 - 정렬 로직을 useSort 훅으로 분리
